requirements.txt
ultralytics>=8.1.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python
matplotlib
pandas
seaborn
tqdm
PyYAML
ipykernel